pyproject.toml
[project]
name = "ajax-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "altgraph==0.17.4",
    "bidict==0.23.1",
    "blinker==1.9.0",
    "click==8.2.1",
    "colorama==0.4.6",
    "faker==37.4.0",
    "flask==3.1.1",
    "flask-cors==6.0.1",
    "flask-session>=0.8.0",
    "flask-socketio==5.5.1",
    "h11==0.16.0",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "markupsafe==3.0.2",
    "mysql-connector-python==9.3.0",
    "packaging==25.0",
    "pefile==2023.2.7",
    "pillow==11.3.0",
    "pyinstaller==6.14.1",
    "pyinstaller-hooks-contrib==2025.5",
    "pymodbus==3.9.2",
    "pyserial==3.5",
    "pystray==0.19.5",
    "python-engineio==4.12.2",
    "python-socketio==5.13.0",
    "pywin32==310",
    "pywin32-ctypes==0.2.3",
    "simple-websocket==1.1.0",
    "six==1.17.0",
    "tzdata==2025.2",
    "werkzeug==3.1.3",
    "wmi==1.5.1",
    "wsproto==1.2.0",
]
